<feat> : 대시보드 관련 버그 수정 및 성능 테스트 완료

+++ back_end/src/main/java/com/example/demo/dashboard/dto/DashboardResponseDto.java
@@ -73,5 +73,5 @@ public class DashboardResponseDto implements Serializable {
         }
         """
     )
-    private Map<String, Integer> domain;
+    private Map<String, Long> domain;
 }
